+++ CHANGES.md
@@ -45,6 +45,8 @@
     - OldCycleMom: Replaced by CycleMom
     - OldOneCycle: Replaced by OneCycle
     - OldLRFinder: Replaced by LRFinder
+    - fold_lr_find: Replaced by lr_find
+    - fold_train_ensemble: Replaced by train_models
 
 ## Comments